Renamed config file and dependencies
