+++ provider/cmd/pulumi-resource-cloud-toolkit-aws/storage/__tests__/fixtures/full-configuration.ts
@@ -1,4 +1,4 @@
-import { BucketVersioningStateArgs } from "../../config";
+import { BucketVersioningStateArgs } from "../../bucketArgs";
 
 export default {
   public: true,
