+++ provider/cmd/pulumi-resource-cloud-toolkit-aws/storage/bucketArgs.ts
